feat: optimise, add the html and js strings to the index directly
